--- a/BlueprintUI/Tests/KeyedTests.swift
+++ b/BlueprintUI/Tests/KeyedTests.swift
@@ -10,11 +10,11 @@ import XCTest
 
 
 class KeyedTests: XCTestCase {
-    
+
     override func setUp() {
         BlueprintView.enableCrossLayoutCaching = true
     }
-    
+
     override func tearDown() {
         BlueprintView.enableCrossLayoutCaching = false
     }