--- a/BlueprintUI/Tests/UIViewElementTests.swift
+++ b/BlueprintUI/Tests/UIViewElementTests.swift
@@ -11,11 +11,11 @@ import XCTest
 
 
 class UIViewElementTests: XCTestCase {
-    
+
     override func setUp() {
         BlueprintView.enableCrossLayoutCaching = true
     }
-    
+
     override func tearDown() {
         BlueprintView.enableCrossLayoutCaching = false
     }