--- a/BlueprintUI/Tests/ViewDiffingBehaviourTests.swift
+++ b/BlueprintUI/Tests/ViewDiffingBehaviourTests.swift
@@ -11,6 +11,15 @@ import XCTest
 @testable import BlueprintUI
 
 class ViewDiffingBehaviourTests: XCTestCase {
+    
+    override func setUp() {
+        BlueprintView.enableCrossLayoutCaching = true
+    }
+    
+    override func tearDown() {
+        BlueprintView.enableCrossLayoutCaching = false
+    }
+
     func test_view_instances_do_not_change() {
         // Set up initial view and contents.