--- a/BlueprintUICommonControls/Tests/Sources/PixelBoundarySnapshotTests.swift
+++ b/BlueprintUICommonControls/Tests/Sources/PixelBoundarySnapshotTests.swift
@@ -3,6 +3,15 @@ import BlueprintUICommonControls
 import XCTest
 
 final class PixelBoundarySnapshotTests: XCTestCase {
+    
+    override func setUp() {
+        BlueprintView.enableCrossLayoutCaching = true
+    }
+    
+    override func tearDown() {
+        BlueprintView.enableCrossLayoutCaching = false
+    }
+
     func test_nestedBoxes() {
         compareSnapshot(
             of: NestedBoxes(addIntermediateViews: false),