--- a/BlueprintUICommonControls/Tests/Sources/TextFieldTests.swift
+++ b/BlueprintUICommonControls/Tests/Sources/TextFieldTests.swift
@@ -4,6 +4,14 @@ import XCTest
 
 
 class TextFieldTests: XCTestCase {
+    
+    override func setUp() {
+        BlueprintView.enableCrossLayoutCaching = true
+    }
+    
+    override func tearDown() {
+        BlueprintView.enableCrossLayoutCaching = false
+    }
 
     func test_snapshots() {