--- a/BlueprintUICommonControls/Tests/Sources/TextFieldTests.swift
+++ b/BlueprintUICommonControls/Tests/Sources/TextFieldTests.swift
@@ -4,11 +4,11 @@ import XCTest
 
 
 class TextFieldTests: XCTestCase {
-    
+
     override func setUp() {
         BlueprintView.enableCrossLayoutCaching = true
     }
-    
+
     override func tearDown() {
         BlueprintView.enableCrossLayoutCaching = false
     }